update users on sync
